Extract remaining pytests from 'references' test module into 'filehashstore' test module and delete redundant 'references' test module
